Extract download to own file
